+++ experimental/packages/exporter-trace-otlp-http/test/node/CollectorTraceExporter.test.ts
@@ -35,6 +35,7 @@ import {
   ensureSpanIsCorrect,
   mockedReadableSpan
 } from '../traceHelper';
+import { nextTick } from 'process';
 import { MockedResponse } from './nodeHelpers';
 import { IExportTraceServiceRequest } from '@opentelemetry/otlp-transformer';
 
@@ -163,7 +164,11 @@ describe('OTLPTraceExporter - node with json over http', () => {
       collectorExporter.export(spans, () => { });
 
       setTimeout(() => {
+        const mockRes = new MockedResponse(200);
         const args = stubRequest.args[0];
+        const callback = args[1];
+        callback(mockRes);
+        mockRes.send('success');
         const options = args[0];
 
         assert.strictEqual(options.hostname, 'foo.bar.com');
@@ -177,7 +182,12 @@ describe('OTLPTraceExporter - node with json over http', () => {
       collectorExporter.export(spans, () => { });
 
       setTimeout(() => {
+        const mockRes = new MockedResponse(200);
         const args = stubRequest.args[0];
+        const callback = args[1];
+        callback(mockRes);
+        mockRes.send('success');
+
         const options = args[0];
         assert.strictEqual(options.headers['foo'], 'bar');
         done();
@@ -188,7 +198,12 @@ describe('OTLPTraceExporter - node with json over http', () => {
       collectorExporter.export(spans, () => { });
 
       setTimeout(() => {
+        const mockRes = new MockedResponse(200);
         const args = stubRequest.args[0];
+        const callback = args[1];
+        callback(mockRes);
+        mockRes.send('success');
+
         const options = args[0];
         assert.strictEqual(options.headers['Content-Encoding'], undefined);
         done();
@@ -199,7 +214,12 @@ describe('OTLPTraceExporter - node with json over http', () => {
       collectorExporter.export(spans, () => { });
 
       setTimeout(() => {
+        const mockRes = new MockedResponse(200);
         const args = stubRequest.args[0];
+        const callback = args[1];
+        callback(mockRes);
+        mockRes.send('success');
+
         const options = args[0];
         const agent = options.agent;
         assert.strictEqual(agent.keepAlive, true);
@@ -209,14 +229,34 @@ describe('OTLPTraceExporter - node with json over http', () => {
     });
 
     it('different http export requests should use the same agent', done => {
-      collectorExporter.export(spans, () => { });
+      const clock = sinon.useFakeTimers();
       collectorExporter.export(spans, () => { });
 
-      setTimeout(() => {
+      const mockRes = new MockedResponse(200);
+      const args = stubRequest.args[0];
+      const callback = args[1];
+
+      callback(mockRes);
+      mockRes.send('success');
+      clock.restore();
+
+      nextTick(() => {
+        const clock = sinon.useFakeTimers();
+        collectorExporter.export(spans, () => { });
+
+        const mockRes2 = new MockedResponse(200);
+        const args2 = stubRequest.args[1];
+        const callback2 = args2[1];
+
+        callback2(mockRes);
+        mockRes2.send('success');
+
         const [firstExportAgent, secondExportAgent] = stubRequest.args.map(
           a => a[0].agent
         );
+
         assert.strictEqual(firstExportAgent, secondExportAgent);
+        clock.restore();
         done();
       });
     });
@@ -242,6 +282,13 @@ describe('OTLPTraceExporter - node with json over http', () => {
       });
 
       collectorExporter.export(spans, () => { });
+
+      const mockRes = new MockedResponse(200);
+      const args = stubRequest.args[0];
+      const callback = args[1];
+
+      callback(mockRes);
+      mockRes.send('success');
     });
 
     it('should log the successful message', done => {
@@ -254,6 +301,7 @@ describe('OTLPTraceExporter - node with json over http', () => {
         const mockRes = new MockedResponse(200);
         const args = stubRequest.args[0];
         const callback = args[1];
+
         callback(mockRes);
         mockRes.send('success');
         setTimeout(() => {
@@ -277,6 +325,7 @@ describe('OTLPTraceExporter - node with json over http', () => {
         const callback = args[1];
         callback(mockResError);
         mockResError.send('failed');
+
         setTimeout(() => {
           const result = responseSpy.args[0][0] as core.ExportResult;
           assert.strictEqual(result.code, core.ExportResultCode.FAILED);
@@ -312,7 +361,6 @@ describe('OTLPTraceExporter - node with json over http', () => {
     });
 
     it('should successfully send the spans', done => {
-      collectorExporter.export(spans, () => { });
       let buff = Buffer.from('');
 
       fakeRequest.on('end', () => {
@@ -325,15 +373,22 @@ describe('OTLPTraceExporter - node with json over http', () => {
 
         ensureExportTraceServiceRequestIsSet(json);
         assert.ok(spySetHeader.calledWith('Content-Encoding', 'gzip'));
-
         done();
       });
 
       fakeRequest.on('data', chunk => {
         buff = Buffer.concat([buff, chunk]);
       });
-    });
 
+      collectorExporter.export(spans,() => { });
+
+      const mockRes = new MockedResponse(200);
+      const args = stubRequest.args[0];
+      const callback = args[1];
+
+      callback(mockRes);
+      mockRes.send('success');
+    });
   });
 
   describe('OTLPTraceExporter - node (getDefaultUrl)', () => {
@@ -356,5 +411,138 @@ describe('OTLPTraceExporter - node with json over http', () => {
         done();
       });
     });
+  });
+  describe('export - with timeout', () => {
+    beforeEach(() => {
+      fakeRequest = new Stream.PassThrough();
+      stubRequest = sinon.stub(http, 'request').returns(fakeRequest as any);
+      spySetHeader = sinon.spy();
+      (fakeRequest as any).setHeader = spySetHeader;
+      (fakeRequest as any).abort = sinon.spy();
+      collectorExporterConfig = {
+        headers: {
+          foo: 'bar',
+        },
+        hostname: 'foo',
+        attributes: {},
+        url: 'http://foo.bar.com',
+        keepAlive: true,
+        httpAgentOptions: { keepAliveMsecs: 2000 },
+        timeoutMillis: 100,
+      };
+      collectorExporter = new OTLPTraceExporter(collectorExporterConfig);
+      spans = [];
+      spans.push(Object.assign({}, mockedReadableSpan));
+    });
+    it('should log the timeout request error message', done => {
+      const responseSpy = sinon.spy();
+      collectorExporter.export(spans, responseSpy);
+
+      setTimeout(() => {
+        fakeRequest.emit('error', { code: 'ECONNRESET'});
+
+        setTimeout(() => {
+          const result = responseSpy.args[0][0] as core.ExportResult;
+          assert.strictEqual(result.code, core.ExportResultCode.FAILED);
+          const error = result.error as OTLPExporterError;
+          assert.ok(error !== undefined);
+          assert.strictEqual(error.message, 'Request Timeout');
+
+          done();
+        });
+      }, 300);
+    });
+  });
+});
+
+describe('export - real http request destroyed before response received', () => {
+  let collectorExporter: OTLPTraceExporter;
+  let collectorExporterConfig: OTLPExporterNodeConfigBase;
+  let spans: ReadableSpan[];
+
+  const server = http.createServer((_, res) => {
+    setTimeout(() => {
+      res.statusCode = 200;
+      res.end();
+    }, 200);
+  });
+  before(done => {
+    server.listen(8081, done);
+  });
+  after(done => {
+    server.close(done);
+  });
+  it('should log the timeout request error message when timeout is 1', done => {
+    collectorExporterConfig = {
+      url: 'http://localhost:8081',
+      timeoutMillis: 1,
+    };
+    collectorExporter = new OTLPTraceExporter(collectorExporterConfig);
+    spans = [];
+    spans.push(Object.assign({}, mockedReadableSpan));
+
+    setTimeout(() => {
+      collectorExporter.export(spans, result => {
+        assert.strictEqual(result.code, core.ExportResultCode.FAILED);
+        const error = result.error as OTLPExporterError;
+        assert.ok(error !== undefined);
+        assert.strictEqual(error.message, 'Request Timeout');
+        done();
+      });
+    }, 0);
+  });
+  it('should log the timeout request error message when timeout is 100', done => {
+    collectorExporterConfig = {
+      url: 'http://localhost:8081',
+      timeoutMillis: 100,
+    };
+    collectorExporter = new OTLPTraceExporter(collectorExporterConfig);
+    spans = [];
+    spans.push(Object.assign({}, mockedReadableSpan));
+
+    setTimeout(() => {
+      collectorExporter.export(spans, result => {
+        assert.strictEqual(result.code, core.ExportResultCode.FAILED);
+        const error = result.error as OTLPExporterError;
+        assert.ok(error !== undefined);
+        assert.strictEqual(error.message, 'Request Timeout');
+        done();
+      });
+    }, 0);
+  });
+});
+
+describe('export - real http request destroyed after response received', () => {
+  let collectorExporter: OTLPTraceExporter;
+  let collectorExporterConfig: OTLPExporterNodeConfigBase;
+  let spans: ReadableSpan[];
+
+  const server = http.createServer((_, res) => {
+    res.write('writing something');
+  });
+  before(done => {
+    server.listen(8081, done);
+  });
+  after(done => {
+    server.close(done);
+  });
+  it('should log the timeout request error message', done => {
+    collectorExporterConfig = {
+      url: 'http://localhost:8081',
+      timeoutMillis: 300,
+    };
+    collectorExporter = new OTLPTraceExporter(collectorExporterConfig);
+    spans = [];
+    spans.push(Object.assign({}, mockedReadableSpan));
+
+    setTimeout(() => {
+      collectorExporter.export(spans, result => {
+        assert.strictEqual(result.code, core.ExportResultCode.FAILED);
+        const error = result.error as OTLPExporterError;
+        assert.ok(error !== undefined);
+        assert.strictEqual(error.message, 'Request Timeout');
+        done();
+      });
+    }, 0);
   });
 });